requirements.txt
# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# ML & Semantic Search
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.0.0
transformers>=4.30.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
alembic>=1.12.1

# Caching
redis>=5.0.0
hiredis>=2.2.3

# Task Queue
celery>=5.3.4
flower>=2.0.1

# File Processing & ETL
pandas>=2.1.3
pyarrow>=14.0.1

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiofiles>=23.2.1

# Legacy support (for backward compatibility)
flask>=3.0.0
flask-cors>=4.0.0
